Renamed base_class_timeseries.py to ts_base_class.py
