+++ ensysmod/database/ts_base_class.py
@@ -4,6 +4,9 @@ from sqlalchemy.sql.schema import ForeignKey
 
 
 class TimeSeriesBase:
+    """
+    Base class for all time series classes.
+    """
     @declared_attr
     def id(self):
         return Column(Integer, primary_key=True)
