+++ ensysmod/model/ts_generation.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, DECIMAL
 
 from ensysmod.database.base_class import Base
-from ensysmod.database.base_class_timeseries import TimeSeriesBase
+from ensysmod.database.ts_base_class import TimeSeriesBase
 
 
 class Generation(TimeSeriesBase, Base):
